+++ python/tests/test_tool_definition_docstring.py
@@ -184,138 +184,6 @@ class TestToolDefinitionDocstring(unittest.TestCase):
         doc = _get_function_docs(func)
 
         assert doc is None
-
-    #my test
-    def sample_func(self, x: int, y: str) -> str:
-        """Sample function to test parsing.
-
-        Args:
-            x (int): The first parameter.
-            y (str): The second parameter.
-
-        Returns:
-            str: The result of the function.
-        """
-        return f"{x} {y}"
-
-    def empty_doc_func(self):
-        """ """
-        pass
-
-    def no_params_func(self) -> int:
-        """Function with no parameters.
-
-        Returns:
-            int: An integer result.
-        """
-        return 0
-
-    def test_get_docstring_view(self):
-        doc_params, doc_returns = get_docstring_view(self.sample_func, DocstringsFormat.GOOGLE, False)
-        self.assertIn("x", doc_params)
-        self.assertEqual(doc_params["x"]["type"], "int")
-        self.assertEqual(doc_params["x"]["required"], True)
-        self.assertIn("y", doc_params)
-        self.assertEqual(doc_params["y"]["type"], "str")
-        self.assertIn("type", doc_returns)
-        self.assertEqual(doc_returns["type"], "str")
-
-        # Test with disable_docstring=True
-        doc_params, doc_returns = get_docstring_view(self.sample_func, DocstringsFormat.GOOGLE, True)
-        self.assertEqual(doc_params, {})
-        self.assertEqual(doc_returns, {})
-
-    def test_parse_function_doc(self):
-        docstring = self.sample_func.__doc__
-        result = _parse_function_doc(docstring)
-        self.assertIn("Sample function to test parsing.", result)
-        self.assertIn("Args:", result)
-
-    def test_parse_function_description_from_docstrings(self):
-        docstring = self.sample_func.__doc__
-        result = _parse_function_description_from_docstrings(docstring)
-        self.assertEqual(result, "Sample function to test parsing.")
-
-        # Test empty docstring
-        self.assertIsNone(_parse_function_description_from_docstrings(""))
-
-    def test_find_and_parse_params_from_docstrings(self):
-        docstring = self.sample_func.__doc__
-        params, returns = _find_and_parse_params_from_docstrings(docstring, DocstringsFormat.GOOGLE)
-        self.assertIn("x", params)
-        self.assertEqual(params["x"]["type"], "int")
-        self.assertEqual(params["x"]["description"], "The first parameter.")
-        self.assertIn("y", params)
-        self.assertEqual(params["y"]["type"], "str")
-        self.assertEqual(params["y"]["description"], "The second parameter.")
-        self.assertEqual(returns["type"], "str")
-        self.assertEqual(returns["description"], "The result of the function.")
-
-        # Test with no parameters or return in docstring
-        empty_params, empty_returns = _find_and_parse_params_from_docstrings(self.no_params_func.__doc__, DocstringsFormat.GOOGLE)
-        self.assertEqual(empty_params, {})
-        self.assertEqual(empty_returns["type"], "int")
-
-    def test_parse_params(self):
-        docstring = self.sample_func.__doc__
-        args_section_start = r"(^|\n)(Args|Arguments|Parameters)\s*:?\s*\n"
-        args_section_end = r"(^|\n)(Returns|Ret)\s*:?\s*\n"
-        param_start = r"(^|\n)\s+(?P<name>[a-zA-Z_][a-zA-Z0-9_]*)\s*(\((?P<type>[^\)]*)\))?\s*:\s*(?=[^\n]+)"
-        params = _parse_params(
-            docstring,
-            args_section_start,
-            args_section_end,
-            param_start,
-            has_name=True
-        )
-        self.assertIn("x", params)
-        self.assertEqual(params["x"]["type"], "int")
-        self.assertEqual(params["x"]["description"], "The first parameter.")
-        self.assertIn("y", params)
-        self.assertEqual(params["y"]["type"], "str")
-        self.assertEqual(params["y"]["description"], "The second parameter.")
-
-        # Test when regex does not match any params
-        empty_params = _parse_params("", args_section_start, args_section_end, param_start, has_name=True)
-        self.assertEqual(empty_params, {})
-
-    def test_docstring_formats(self):
-        # Test different formats
-        numpy_docstring = """Function with Numpy format.
-
-        Parameters
-        ----------
-        x : int
-            The first parameter.
-        y : str
-            The second parameter.
-
-        Returns
-        -------
-        str
-            The result of the function.
-        """
-        params, returns = _find_and_parse_params_from_docstrings(numpy_docstring, DocstringsFormat.NUMPY)
-        self.assertIn("x", params)
-        self.assertEqual(params["x"]["type"], "int")
-        self.assertEqual(params["x"]["description"], "The first parameter.")
-        self.assertEqual(returns["type"], "str")
-
-        # Sphinx format example
-        sphinx_docstring = """Function with Sphinx format.
-
-        :param x: The first parameter.
-        :type x: int
-        :param y: The second parameter.
-        :type y: str
-        :returns: The result of the function.
-        :rtype: str
-        """
-        params, returns = _find_and_parse_params_from_docstrings(sphinx_docstring, DocstringsFormat.SPHINX)
-        self.assertIn("x", params)
-        self.assertEqual(params["x"]["type"], "int")
-        self.assertEqual(params["x"]["description"], "The first parameter.")
-        self.assertEqual(returns["type"], "str")
-
+        
 if __name__ == '__main__':
     unittest.main()
